requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
mysql-connector-python==9.0.0
numpy==1.26.4